moved options classes to separate folder
