+++ backend/Tasque.BLL/JWT/JwtFactory.cs
@@ -2,6 +2,7 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
+using Tasque.Core.BLL.Options;
 
 namespace Tasque.Core.BLL.JWT
 {
